Combine RangeSelectorInner into RangeSelector
